Extracts pagination listener into reusable class
