+++ src/main/java/com/keenwrite/typesetting/Typesetter.java
@@ -4,6 +4,8 @@ package com.keenwrite.typesetting;
 import com.keenwrite.util.GenericBuilder;
 
 import java.nio.file.Path;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.concurrent.Callable;
 
 /**
@@ -19,56 +21,6 @@ public class Typesetter {
 
   public static GenericBuilder<Mutator, Typesetter> builder() {
     return GenericBuilder.of( Mutator::new, Typesetter::new );
-  }
-
-  private final Mutator mMutator;
-
-  /**
-   * Creates a new {@link Typesetter} instance capable of configuring the
-   * typesetter used to generate a typeset document.
-   */
-  Typesetter( final Mutator mutator ) {
-    assert mutator != null;
-
-    mMutator = mutator;
-  }
-
-  public void typeset() throws Exception {
-    final Callable<Void> typesetter;
-
-    if( HostTypesetter.isReady() ) {
-      typesetter = new HostTypesetter( mMutator );
-    }
-    else if( GuestTypesetter.isReady() ) {
-      typesetter = new GuestTypesetter( mMutator );
-    }
-    else {
-      throw new TypesetterNotFoundException( TYPESETTER_EXE );
-    }
-
-    typesetter.call();
-  }
-
-  protected Path getInputPath() {
-    return mMutator.getInputPath();
-  }
-
-  protected Path getOutputPath() {
-    return mMutator.getOutputPath();
-  }
-
-  protected Path getThemePath() {
-    return mMutator.getThemePath();
-  }
-
-  /**
-   * Answers whether logs and other files should be deleted upon error. The
-   * log files are useful for debugging.
-   *
-   * @return {@code true} to delete generated files.
-   */
-  public boolean autoclean() {
-    return mMutator.isAutoClean();
   }
 
   public static final class Mutator {
@@ -124,6 +76,80 @@ public class Typesetter {
     }
   }
 
+  private final Mutator mMutator;
+
+  /**
+   * Creates a new {@link Typesetter} instance capable of configuring the
+   * typesetter used to generate a typeset document.
+   */
+  Typesetter( final Mutator mutator ) {
+    assert mutator != null;
+
+    mMutator = mutator;
+  }
+
+  public void typeset() throws Exception {
+    final Callable<Void> typesetter;
+
+    if( HostTypesetter.isReady() ) {
+      typesetter = new HostTypesetter( mMutator );
+    }
+    else if( GuestTypesetter.isReady() ) {
+      typesetter = new GuestTypesetter( mMutator );
+    }
+    else {
+      throw new TypesetterNotFoundException( TYPESETTER_EXE );
+    }
+
+    typesetter.call();
+  }
+
+  /**
+   * Generates the command-line arguments used to invoke the typesetter.
+   */
+  List<String> options() {
+    final var themePath = getThemePath();
+    final var inputDoc = getInputPath().toString();
+    final var outputDoc = getOutputPath().getFileName();
+
+    final var args = new LinkedList<String>();
+
+    args.add( "--autogenerate" );
+    args.add( "--script" );
+    args.add( "mtx-context" );
+    args.add( "--batchmode" );
+    args.add( "--nonstopmode" );
+    args.add( "--purgeall" );
+    args.add( "--path='" + themePath + "'" );
+    args.add( "--environment='main'" );
+    args.add( "--result='" + outputDoc + "'" );
+    args.add( inputDoc );
+
+    return args;
+  }
+
+  protected Path getInputPath() {
+    return mMutator.getInputPath();
+  }
+
+  protected Path getOutputPath() {
+    return mMutator.getOutputPath();
+  }
+
+  private Path getThemePath() {
+    return mMutator.getThemePath();
+  }
+
+  /**
+   * Answers whether logs and other files should be deleted upon error. The
+   * log files are useful for debugging.
+   *
+   * @return {@code true} to delete generated files.
+   */
+  public boolean autoclean() {
+    return mMutator.isAutoClean();
+  }
+
   public static boolean canRun() {
     return hostCanRun() || guestCanRun();
   }
@@ -135,5 +161,4 @@ public class Typesetter {
   private static boolean guestCanRun() {
     return GuestTypesetter.isReady();
   }
-
 }
